cmd/sprig: fix environment variable parsing
`os.Environ()` returns a slice of "key=value" strings representing the
environment. This commit fixes the behavior of sprigCommand.vals() to
parse environment variables correctly. It also renames `envKeys` to
`envVars` as it is slightly more descriptive of the contents and adds a
test to confirm that environment variables are read correctly.